requirements.txt
Flask>=3.1.1
pdfminer.six>=20250506
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=9.0.0
gunicorn>=20.1.0
